set server route on client initialization
